pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eaam"
version = "0.1.0"
description = "Emotion-Anchored Associative Memory — biologically-inspired memory for AI"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["ai", "memory", "associative", "emotion", "llm", "mcp"]
dependencies = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
eaam = "eaam.cli:main"

[project.urls]
Homepage = "https://github.com/XoAnonXo/eaam"
Repository = "https://github.com/XoAnonXo/eaam"

[tool.hatch.build.targets.wheel]
packages = ["eaam"]

[tool.hatch.build.targets.wheel.force-include]
"eaam/engine/surfaces.yaml" = "eaam/engine/surfaces.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
